+++ io/bitstream.js
@@ -20,7 +20,7 @@ const BITMASK = [0, 0x01, 0x03, 0x07, 0x0F, 0x1F, 0x3F, 0x7F, 0xFF];
  *
  * Bit reading always proceeds from the first byte in the buffer, to the
  * second byte, and so on. The MTL flag controls which bit is considered
- * first *inside* the byte.
+ * first *inside* the byte. The default is least-to-most direction.
  *
  * An Example for how Most-To-Least vs Least-to-Most mode works:
  *